Code/sakahan-backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.37.8
botocore==1.37.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==24.3
django-ses==4.4.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docopt==0.6.2
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
pyjsparser==2.7.1
PyJWT==2.9.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
s3transfer==0.11.4
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
